TelegramRegisterBot/requirements.txt
aiofiles==0.7.0
aiohttp==3.9.5
aiohttp-sse-client==0.2.1
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.3.0
asgiref==3.4.1
astroid==2.3.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.1
bitarray==2.9.2
blinker==1.8.2
brute==0.0.3
bs4==0.0.1
certifi==2022.9.24
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.7
click-default-group==1.2.2
cloup==0.7.1
cmake==3.26.3
colorama==0.4.6
colour==0.1.5
commonmark==0.9.1
contourpy==1.1.1
cycler==0.12.1
db0mb3r==3.1.4.17
decorator==5.1.0
discord.py==2.0.0
docopt==0.6.2
docutils==0.19
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.89.1
filelock==3.9.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
glcontext==2.3.4
greenlet==3.0.3
h11==0.14.0
httpcore==0.13.6
httpx==0.19.0
huggingface-hub==0.22.2
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
install==1.3.4
isort==4.3.21
isosurfaces==0.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.5
lazy-object-proxy==1.4.3
loguru==0.5.3
lxml==4.9.1
mapbox-earcut==0.12.10
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.6.1
moderngl==5.6.4
moderngl-window==2.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==6.1.0
multidict==6.0.5
multipledispatch==0.6.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==1.5.3
pep517==0.12.0
phonenumbers==8.12.31
pillow==10.3.0
pip-tools==6.4.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
prometheus_client==0.20.0
PyAutoGUI==0.9.52
pybind11==2.10.3
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==1.10.15
pydnevnikruapi==0.1.1
pydub==0.25.1
PyGetWindow==0.0.9
pyglet==1.5.21
Pygments==2.10.0
pylint==2.4.4
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyRect==0.1.4
pyrr==0.10.3
PyScreeze==0.1.27
PySocks==1.7.1
pyTelegramBotAPI==4.7.0
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-imagesearch==1.2.0
python3-xlib==0.15
pytoniq-core==0.1.36
PyTweening==1.0.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
random-user-agent==1.0.1
referencing==0.35.1
regex==2024.4.16
requests==2.31.0
rfc3986==1.5.0
rich==10.11.0
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.0.2
scipy==1.10.0
screeninfo==0.6.7
setuptools-scm==7.0.5
six==1.16.0
skia-pathops==0.7.0
sniffio==1.3.1
soupsieve==2.3.2.post1
spark-parser==1.8.9
SQLAlchemy==2.0.30
starlette==0.22.0
sympy==1.12
telebot==0.0.5
threadpoolctl==3.4.0
tokenizers==0.19.1
tomli==1.2.1
ton-client-py==1.44.3.0
torch==2.2.2
tqdm==4.64.1
typing_extensions==4.11.0
uncompyle==2.0.0
uncompyle6==3.8.0
unrar==0.4
urllib3==1.26.12
uvicorn==0.18.3
validators==0.18.2
watchdog==2.1.6
Werkzeug==3.0.3
win32-setctime==1.0.3
wincertstore==0.2
wrapt==1.11.2
x25519==0.0.2
xdis==6.0.3
yarg==0.1.9
yarl==1.9.4
zipp==3.18.1
